fix
git-svn-id: http://svn.eurecom.fr/openair4G/trunk@7509 818b1a75-f10b-46b9-bf7c-635c3b92a50f

+++ openair1/PHY/TOOLS/time_meas.c
@@ -67,7 +67,7 @@ void print_meas_now(time_stats_t *ts, const char* name, int subframe, FILE* file
     if (ts->trials>0) {
 
       //fprintf(file_name,"Name %25s: Processing %15.3f ms for SF %d, diff_now %15.3f \n", name,(ts->diff_now/(cpu_freq_GHz*1000000.0)),subframe,ts->diff_now);
-      fprintf(file_name,"%15.3f ms, diff_now %15.3f \n",(ts->diff_now/(cpu_freq_GHz*1000000.0)),ts->diff_now);
+      fprintf(file_name,"%15.3f ms, diff_now %15.3f \n",(ts->diff_now/(cpu_freq_GHz*1000000.0)),(double)ts->diff_now);
       
     }
   }
